refactor is now properly checkpointing.
